Add more ListManager tests
